icp_fusion/requirements.txt
numpy
matplotlib
open3d
numpy-quaternion
